refactor: 상수를 폴더로 구분하여 구조 개선
