+++ src/constants/errors/prefix.js
@@ -1,0 +1,1 @@
+export const ERROR_PREFIX = "[ERROR]";
