+++ src/constants/rules/car.js
@@ -1,0 +1,5 @@
+export const CAR_RULES = Object.freeze({
+  MIN_CAR_NAME_LENGTH: 1,
+  MAX_CAR_NAME_LENGTH: 5,
+  CAR_NAME_SEPARATOR: ",",
+});
