+++ src/constants/rules/randomNumber.js
@@ -1,0 +1,4 @@
+export const RANDOM_RULES = Object.freeze({
+  MIN_RANDOM_NUMBER: 0,
+  MAX_RANDOM_NUMBER: 9,
+});
